Fix bug when indexing StateMonitor with `[:3]` type idnexing
Fixes #50. Issue was the same as in PR brian-team/brian2#1119. We need
to update the monitor's `N` variable on CUDA side and we need to update
the device variable since it is copied to the host variable in
`write_arrays()`. And we need to use `WRITES_TO_READ_ONLY_VARIABLES`,
otherwise `N` will be cached in brian2 and not loaded after simulation.