Remove check for occurrences of multiple pre-post synapses
This check is only required when one of the `SynapticPathway` objects of
the `Synapses` object does not use atomics and has synaptic effect mode
`target`. Since we use atomics by default now, I removed the check. If
we ever want to be able to not use atomics, we need to reactivate the
multi occurrence check here accordingly (see comments in code of this
commit).